[FLINK-24807][iteration] not emit CoordinatorCheckpointEvent after terminating
Currently the HeadCoordinator would emit CoordinatorCheckpointEvent to
the tasks so that the GloballyAlignedEvent would not be interleave with
the checkpoint barrier. Howver, if the tasks are finished and we
continue emitting the event, the checkpoint would fail due to there
are failed operator events. To address this issue, we would stop
emitting CoordinatorCheckpointEvent after the head operator is terminating,
namely it received the GloballyAlignedEvent marking terminating.